+++ ctp2_code/ui/aui_common/aui_hypertextbox.cpp
@@ -255,7 +255,7 @@ AUI_ERRCODE aui_HyperTextBox::AddHyperStatics( const MBCHAR *hyperText )
 			{
 				aui_Static *hs = CreateHyperStatic(
 					ptr,
-					len,
+					static_cast<uint32>(len),
 					m_hyperTtffile,
 					m_hyperPointSize,
 					m_hyperBold,
@@ -340,7 +340,7 @@ AUI_ERRCODE aui_HyperTextBox::AddHyperStatics( const MBCHAR *hyperText )
 						else
 						{
 							len -= truncLen;
-							hs->SetText( start, truncLen );
+							hs->SetText( start, static_cast<uint32>(truncLen) );
 
 							hs->Move( m_curStaticPos.x, m_curStaticPos.y );
 							hs->Resize( penPos.x, hs->Height() );
